Fixes #22: Invoke `init()`s on `document` rather than `window` ready.
This addresses an issue that is *probably* caused by a race condition in
Chrome and Tampermonkey as described here:
https://forum.tampermonkey.net/viewtopic.php?t=1135

The effect of the bug is that, in Chrome, the script's `init()` methods
were never being called. Using jQuery to ensure that they get called
when the DOM is ready (seems to have) fixed the issue.